RecruitApp2/RecruitAppBackend/requirements.txt
Flask==2.1.0
azure-cosmos==4.2.0
